# docs/es/propuesta_motor/app/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API para el Proof of Concept del Motor Contable Colombiano",
        "title": "Motor Contable POC API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Soporte API",
            "email": "soporte@motorcontable.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/organization/current": {
            "get": {
                "description": "Retorna la información completa de la organización actual del sistema",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Organización"
                ],
                "summary": "Obtiene organización actual",
                "responses": {
                    "200": {
                        "description": "Información de la organización",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.StandardResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.OrganizationDetail"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Organización no encontrada",
                        "schema": {
                            "$ref": "#/definitions/models.StandardResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/models.StandardResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Actualiza la configuración de la organización actual",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Organización"
                ],
                "summary": "Actualiza organización",
                "parameters": [
                    {
                        "description": "Datos de la organización a actualizar",
                        "name": "organization",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Organización actualizada",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.StandardResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.OrganizationDetail"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Datos inválidos",
                        "schema": {
                            "$ref": "#/definitions/models.StandardResponse"
                        }
                    },
                    "404": {
                        "description": "Organización no encontrada",
                        "schema": {
                            "$ref": "#/definitions/models.StandardResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/models.StandardResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/organization/dashboard": {
            "get": {
                "description": "Retorna métricas y KPIs principales de la organización",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Organización"
                ],
                "summary": "Dashboard de organización",
                "responses": {
                    "200": {
                        "description": "Datos del dashboard",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.StandardResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Organización no encontrada",
                        "schema": {
                            "$ref": "#/definitions/models.StandardResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/models.StandardResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/organization/validate": {
            "post": {
                "description": "Ejecuta validaciones de reglas contables y de negocio sobre la organización",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Organización"
                ],
                "summary": "Valida reglas de negocio",
                "responses": {
                    "200": {
                        "description": "Resultado de validaciones",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.StandardResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Organización no encontrada",
                        "schema": {
                            "$ref": "#/definitions/models.StandardResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/models.StandardResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.AccountingConfig": {
            "type": "object",
            "properties": {
                "allow_negative_inventory": {
                    "type": "boolean"
                },
                "auto_numbering": {
                    "type": "boolean"
                },
                "current_period": {
                    "type": "string"
                },
                "default_payment_terms": {
                    "type": "integer"
                },
                "default_tax_rate": {
                    "type": "number"
                },
                "last_closed_period": {
                    "type": "string"
                },
                "require_balanced_entries": {
                    "type": "boolean"
                }
            }
        },
        "models.ContactInfo": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "postal_code": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "website": {
                    "type": "string"
                }
            }
        },
        "models.DSLConfig": {
            "type": "object",
            "properties": {
                "auto_process_vouchers": {
                    "type": "boolean"
                },
                "error_handling": {
                    "type": "string"
                },
                "parallel_processing": {
                    "type": "boolean"
                },
                "template_cache_enabled": {
                    "type": "boolean"
                },
                "validation_level": {
                    "type": "string"
                }
            }
        },
...
        "models.ErrorInfo": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "models.FiscalInfo": {
            "type": "object",
            "properties": {
                "accounting_standard": {
                    "type": "string"
                },
                "decimal_places": {
                    "type": "integer"
                },
                "fiscal_year_end": {
                    "type": "string"
                },
                "fiscal_year_start": {
                    "type": "string"
                },
                "reporting_currency": {
                    "type": "string"
                },
                "rounding_method": {
                    "type": "string"
                },
                "tax_regime": {
                    "type": "string"
                }
            }
        },
        "models.OrganizationDetail": {
            "type": "object",
            "properties": {
                "accounting_configuration": {
                    "$ref": "#/definitions/models.AccountingConfig"
                },
                "code": {
                    "type": "string"
                },
                "commercial_name": {
                    "type": "string"
                },
                "contact_info": {
                    "$ref": "#/definitions/models.ContactInfo"
                },
                "country_code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "currency_default": {
                    "type": "string"
                },
                "document_type": {
                    "type": "string"
                },
                "dsl_configuration": {
                    "$ref": "#/definitions/models.DSLConfig"
                },
                "fiscal_info": {
                    "$ref": "#/definitions/models.FiscalInfo"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "language": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "statistics": {
                    "$ref": "#/definitions/models.OrganizationStatistics"
                },
                "subscription_info": {
                    "$ref": "#/definitions/models.SubscriptionInfo"
                },
                "system_settings": {
                    "$ref": "#/definitions/models.SystemSettings"
                },
                "tax_id": {
                    "type": "string"
                },
                "timezone": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_permissions": {
                    "$ref": "#/definitions/models.UserPermissions"
                }
            }
        },
        "models.OrganizationStatistics": {
            "type": "object",
            "properties": {
                "active_users": {
                    "type": "integer"
                },
                "api_calls_month": {
                    "type": "integer"
                },
                "storage_used_mb": {
                    "type": "integer"
                },
                "total_accounts": {
                    "type": "integer"
                },
                "total_journal_entries": {
                    "type": "integer"
                },
                "total_vouchers": {
                    "type": "integer"
                }
            }
        },
...
        "models.StandardResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/models.ErrorInfo"
                },
                "success": {
                    "type": "boolean"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "models.SubscriptionInfo": {
            "type": "object",
            "properties": {
                "current_month_vouchers": {
                    "type": "integer"
                },
                "current_users": {
                    "type": "integer"
                },
                "expires_at": {
                    "type": "string"
                },
                "features": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "max_monthly_vouchers": {
                    "type": "integer"
                },
                "max_users": {
                    "type": "integer"
                },
                "plan": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "models.SystemSettings": {
            "type": "object",
            "properties": {
                "audit_level": {
                    "type": "string"
                },
                "backup_frequency": {
                    "type": "string"
                },
                "notification_preferences": {
                    "type": "object",
                    "additionalProperties": true
                },
                "retention_period_days": {
                    "type": "integer"
                }
            }
        },
...
        "models.UserPermissions": {
            "type": "object",
            "properties": {
                "can_close_periods": {
                    "type": "boolean"
                },
                "can_edit_dsl_templates": {
                    "type": "boolean"
                },
                "can_edit_organization": {
                    "type": "boolean"
                },
                "can_manage_users": {
                    "type": "boolean"
                },
                "can_view_all_data": {
                    "type": "boolean"
                }
            }
        },
...
    }
}